test(personal-assistant): run owner permission matrix by default
