+++ plugins/plugin-personal-assistant/package.json
@@ -50,7 +50,7 @@
     "pretest": "node scripts/lint-default-packs.mjs",
     "test": "vitest run --config vitest.config.ts",
     "test:background-real": "vitest run --config vitest.background-real.config.ts test/scheduled-task-end-to-end.e2e.test.ts test/reminder-review-job.real.e2e.test.ts test/lifeops-scheduling.real.test.ts test/schedule-merged-state.real.test.ts",
-    "test:integration": "vitest run --config vitest.src-integration.config.ts && cd \"$(git rev-parse --show-toplevel)\" && vitest run --config packages/scripts/vitest/integration.config.ts plugins/plugin-personal-assistant/test/integration-lane.smoke.integration.test.ts plugins/plugin-personal-assistant/test/life-smoke.integration.test.ts plugins/plugin-personal-assistant/test/owner-agent-permission-matrix.integration.test.ts",
+    "test:integration": "vitest run --config vitest.src-integration.config.ts && cd \"$(git rev-parse --show-toplevel)\" && vitest run --config packages/scripts/vitest/integration.config.ts plugins/plugin-personal-assistant/test/integration-lane.smoke.integration.test.ts plugins/plugin-personal-assistant/test/life-smoke.integration.test.ts",
     "test:scenarios": "cd ../.. && SCENARIO_USE_DETERMINISTIC_MODEL=1 bun packages/scenario-runner/src/cli.ts run packages/test/scenarios/reminders --lane pr-deterministic && cd plugins/plugin-personal-assistant && bun run test:scenarios:lifeops-spine",
     "test:scenarios:lifeops-spine": "cd ../.. && SCENARIO_USE_DETERMINISTIC_MODEL=1 bun packages/scenario-runner/src/cli.ts run packages/scenario-runner/test/scenarios --lane pr-deterministic --scenario deterministic-lifeops-scheduled-tasks,deterministic-lifeops-dispatch-retry,deterministic-lifeops-recurrence,deterministic-lifeops-concurrent-day,deterministic-lifeops-multiday-journey",
     "test:scenarios:list": "cd ../.. && bun packages/scenario-runner/src/cli.ts list plugins/plugin-personal-assistant/test/scenarios",
